split related func into own files
